Casting indexLast to bigint
ROW_NUMBER returns a bigint. When comparing to an integer I have found that SQL Server will sometime convert the bigint to an int on every row in the result. For larger tables this can slow the query down substantially. By converting indexLast to a bigint SQL Server does not have to convert anything.